+++ src/Inline/Parser/CloseBracketParser.php
@@ -162,7 +162,7 @@ class CloseBracketParser implements InlineParserInterface, EnvironmentAwareInter
         $title = '';
         // make sure there's a space before the title:
         if (\preg_match(RegexHelper::REGEX_WHITESPACE_CHAR, $cursor->peek(-1))) {
-            $title = LinkParserHelper::parseLinkTitle($cursor) ?: '';
+            $title = LinkParserHelper::parseLinkTitle($cursor) ?? '';
         }
 
         $cursor->advanceToNextNonSpaceOrNewline();